Complete Phase 1-3: bug fixes, integration, polish
Phase 1 - Bug Fixes:
- Fix Transform::forward() to return -Z (right-handed convention)
- Fix PhysicsWorld query pipeline caching (lazy rebuild)
- Add AssetStore path dedup, generation checks, count()
- Fix App system ordering (deterministic stage iteration)
- Add World::despawn() with full component cleanup

Phase 2 - Integration:
- Implement scene save/load (JSON serialization via serde)
- Add OBJ mesh loader with fan triangulation
- Per-entity material override in renderer (optional bind groups)
- Live editor Transform + MaterialOverride editing (egui inspector)
- Per-entity Lua scripting (behaviour tables: on_init/on_update)
- Render batching by mesh + frustum culling (AABB/Frustum)
- Rewrite examples to use run() API (75%+ code reduction)

Phase 3 - Polish:
- Spatial audio (inverse-distance attenuation + stereo panning)
- Input action mapping (ActionMap with multi-binding + axis)
- Expanded engine prelude with all new types
- Fix all clippy warnings (map_or, div_ceil, redundant closures)

+++ crates/quasar-scripting/src/bridge.rs
@@ -155,7 +155,7 @@ pub fn register_bridge(lua: &Lua) -> LuaResult<()> {
     fn push_command(lua: &Lua, cmd: LuaTable) -> LuaResult<()> {
         let quasar: LuaTable = lua.globals().get("quasar")?;
         let commands: LuaTable = quasar.get("_commands")?;
-        let len = commands.len()? as i64;
+        let len = commands.len()?;
         commands.set(len + 1, cmd)?;
         Ok(())
     }
